refactor(router): refactor InstructionResolver
